Remake a modconfig in Functions.
