+++ Sources/Everglow.Function/FeatureFlags/EverglowConfig.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel;
-using Everglow.Commons.AssetReplace.UIReplace.Core;
 using Terraria.ModLoader.Config;
 
 namespace Everglow.Commons.FeatureFlags;
@@ -27,64 +26,4 @@ public class EverglowConfig : ModConfig
 			return ModContent.GetInstance<EverglowConfig>().debugMode;
 		}
 	}
-}
-
-// 以后解决这个Config
-public class EverglowClientConfig : ModConfig
-{
-	/// <summary>
-	/// 用于各个客户端的个性化设置
-	/// </summary>
-	public override ConfigScope Mode => ConfigScope.ClientSide;
-
-	[Header("TextureReplace")]
-	[DefaultValue(TextureReplaceMode.Terraria)]
-	[DrawTicks]
-	public TextureReplaceMode TextureReplace;
-
-	[Header("AudioReplace")]
-
-	[DefaultValue(true)]
-	public bool ItemPickSoundReplace;
-
-	[DefaultValue(TuskAudioReplaceMode.TuskFighting)]
-	[DrawTicks]
-	public TuskAudioReplaceMode TuskAudioReplace;
-
-	public override void OnChanged()
-	{
-		if ((int)TextureReplace >= 5)
-			TextureReplace = TextureReplaceMode.Terraria;
-		if ((int)TuskAudioReplace >= 2)
-			TuskAudioReplace = TuskAudioReplaceMode.TuskFighting;
-		if (UIReplaceModule.IsLoaded)
-		{
-			UIReplaceModule.ReplaceTextures(TextureReplace);
-		}
-
-		base.OnChanged();
-	}
-
-	public static int ReplaceTuskAudio
-	{
-		get
-		{
-			return (int)ModContent.GetInstance<EverglowClientConfig>().TuskAudioReplace;
-		}
-	}
-}
-
-public enum TextureReplaceMode
-{
-	Terraria,
-	Default,
-	EternalResolve,
-	Everglow,
-	Myth,
-}
-
-public enum TuskAudioReplaceMode
-{
-	TuskFighting,
-	OldTuskFighting,
 }
